+++ part-3/finish/Metal Workshop/PulsingCircleShaders.metal
@@ -34,7 +34,7 @@ vertex OutVertex pulsing_circle_vertex_main(constant InVertex *vertices [[buffer
 	return outVertex;
 };
 
-fragment half4 pulsing_circle_fragment_main(OutVertex outVertex [[stage_in]])
+fragment float4 pulsing_circle_fragment_main(OutVertex outVertex [[stage_in]])
 {
 	//calculate distance from origin
 	float2 flippedWindowOrigin = float2(outVertex.windowOrigin[0],
@@ -48,8 +48,8 @@ fragment half4 pulsing_circle_fragment_main(OutVertex outVertex [[stage_in]])
 	//get circle radius
 	float circleRadius = min(outVertex.windowSize[0], outVertex.windowSize[1])/4;
 	if (distance < circleRadius) {
-		return half4(outVertex.color);
+		return float4(outVertex.color);
 	} else {
-		return half4(0);
+		return float4(0);
 	}
 }
